requirements.txt
flask>=3.0
flask-cors
textblob
vaderSentiment
requests
python-dotenv
gunicorn
beautifulsoup4
lxml
